changes json constructor from int to long

+++ post-4/src/main/java/io/vertx/intro/first/Article.java
@@ -28,7 +28,7 @@ public class Article {
 
     public Article(JsonObject json) {
         this(
-            json.getInteger("id", -1),
+            json.getLong("id", -1l),
             json.getString("title"),
             json.getString("url")
         );
